Code didn't show up before.
I copied and pasted it in from Visual Studios. I couldn't get the file to show up. I was fiddling around trying to make it go on a different repository but that didn't work. I'm not sure if this will work. But, I've copied and pasted all of my work so hopefully it does what it's supposed to.